[PBW-3944] Video is getting replayed after clicking on skip button for Postroll Ad
Prevents Android from replaying unless there are 2 click events at the
end of an ad. This is because Android passes a skip-ad click to the
main video: “Whenever the user clicks on skip ad (pre/mid/post) in
order to resume the main video it generates a click on the screen which
triggers the togglePlayPause() function. Hence in case of post roll ads
the click on skip ad is causing the click on end screen which in turn
replays the video.”