[PBW-4199] Overlay close button opening ad pages
In the event listener handleOverlayClick() (adOverlay.js), adds another
condition to check if the click event target is not a button(close
button).
Furthermore in playingScreen.js within handlePlayerMouseUp() event
listener adds the same condition to check if the click was not
originated from the button(close button of overlay).