[PBW-4390] CC button staying on replay
Ensure that the available languages is null when setting a new video.
We cannot do this in ‘Played’ because we cannot be sure a prior video
that has CC was fully played before a new video is set, so nulling the
list on setting an embed code, much as is done with availableBitrates.