[PBW-4542] We are unable to perform pause/play after closing overlay.
Even though overlay is being closed, the plugin object is still present
in front of the UI layer of the skin in order to fix a separate PBI
issue that was allowing users to skip past ads. This fix keeps the
object, but makes it take up no screen space, so that the UI can be
interacted with.